working on parsing a big file
issues with handling my JSON data, and also the header and all that.

works okay right now, test isn't done though

also, issues with CSV not dumping quotes??  haven't tested for those
cases when absolutely necessary.

Also also, not sure if I'm handling "header" correctly when it is just
stripped off of the JSON object like this case is doing.